feat: cloud functions support maybe
