+++ deploy/create-indexes.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+cd `dirname $0`
+gcloud datastore indexes create ../index.yaml
